decoder: fix attribute snippet generation for objects
All attributes require equal sign (=).

Terraform's parser in particular can parse Set(Object),
or List(Object) *attribute* also as a block.

For example vault_auth_backend's tune attribute can be written as

tune = [ {
  max_lease_ttl      = "90000s"
  listing_visibility = "unauth"
} ]

or also as

tune {
  max_lease_ttl      = "90000s"
  listing_visibility = "unauth"
}

The first form (*with* "= []") is preferred, so
this is reflected in snippets.

The decoder currently doesn't understand the block form.
That is something we may need to port from Terraform core's
lang/blocktoattr package once we need it e.g. for completion
inside such attribute or for hover.