Collect and match `ImpliedOrigin` during reference origin collection
We match the implied origins from the schema against each
`LocalOrigin` we found. If the address matches, we introduce a
new `PathOrigin`, so the user can navigate from this origin to the
target of the implied origin.

This is primarily used for navigating to output block of modules
in Terraform.