decoder: Fix return type comparison in functionExpr
go-cty's understanding of type conformity is a lot stricter than Terraform's, mostly for historical reasons.

For example, in Terraform `"42"` is convertible to `42`.
Similarly, in `number_attr = "${var.number}"` the template expression is technically a string,
and while this is *not* the preferred way to write it, it works nonetheless.
This used to be the only way to do interpolation in older Terraform versions (<0.12).

Some other products (language servers) may find the stricter comparison more helpful, but we don't have
enough real feedback/use cases to make the comparison customizable in any way. It is something we could
consider in the future though.