decoder: Collect interpolated targets for Any
In a perfect/ideal scenario we could collect just targets which match the type
but in reality we cannot always establish the expression type, because:
 - we don't recognise all expression types (yet)
 - we'd have to re-implement DAG in some way to ensure we collect and interpolate in the right order
 - not all variables may be known at editing time (some may be provided only at runtime)

Also it's questionable whether it's beneficial to suppress mismatching targets.
We _mostly_ only do it as a hack to aid LiteralType and Reference in OneOf.