Support schema extensions in early validation
This does not raise diagnostics if schema extensions are enabled and Count or ForEach is enabled.

In the work on functions & expressions we introduced the concept of extensions for handling count and for_each. Both aren't part of the regular attributes for a block, but are introduced by enabling the specific extensions for a block.

This adds count and for_each at runtime like completion, hover, and other parts do. Since these are optional attributes, there is are no diagnostics necessary if they are not present.