Introduce `SupportUnknownNestedRefs` for block addresses
This will allow references for nested origins that point to a block that
doesn't have the referenced attributes and doesn't point to a nested
target (like a module).

This is particularly useful for store blocks in the Stack language that
can point to a varset defined in HCP Terraform. With the new flag, we
can have valid references without knowing the varset entries.